+++ R/fit.R
@@ -8,6 +8,12 @@
 #' @inheritParams mmrm
 #' @param control (`mmrm_control`)\cr object.
 #' @param ... Additional arguments to pass to [mmrm_control()].
+#'
+#' @details
+#' `fit_single_optimizer` will fit the `mmrm` model using the `control` provided.
+#' If there are mutliple optimizers provided in `control`, only the first optimizer
+#' will be used.
+#'
 #' @return The `mmrm_fit` object, with additional attributes containing warnings,
 #'   messages, optimizer used and convergence status in addition to the
 #'   `mmrm_tmb` contents.
@@ -32,6 +38,7 @@ fit_single_optimizer <- function(formula,
   assert_vector(weights)
   assert_logical(reml)
   assert_class(control, "mmrm_control")
+  assert_list(control$optimizers, names = "unique")
   quiet_fit <- h_record_all_output(
     fit_mmrm(
       formula = formula,
@@ -194,8 +201,8 @@ mmrm_control <- function(n_cores = 1L,
   assert_character(method)
   assert_numeric(start, null.ok = TRUE)
   assert_flag(accept_singular)
+  assert_list(optimizers, names = "unique")
   method <- match.arg(method)
-
   structure(
     list(
       optimizers = optimizers,
@@ -240,7 +247,7 @@ mmrm_control <- function(n_cores = 1L,
 #'
 #' When optimizer is not set, first the default optimizer
 #' (`L-BFGS-B`) is used to fit the model. If that converges, this is returned.
-#' If not, the other available optimizers from [refit_multiple_optimizers()] are
+#' If not, the other available optimizers, including `BFGS`, `CG` and `nlminb` are
 #' tried (in parallel if `n_cores` is set and not on Windows).
 #' If none of the optimizers converge, then the function fails. Otherwise
 #' the best fit is returned.