+++ R/kenwardroger.R
@@ -42,7 +42,7 @@ h_get_kr_comp <- function(tmb_data, theta) {
 #' @keywords internal
 h_df_md_kr <- function(object, contrast, linear = FALSE) {
   assert_class(object, "mmrm")
-  assert_logical(linear)
+  assert_flag(linear)
   assert_matrix(contrast, mode = "numeric", any.missing = FALSE, ncols = length(component(object, "beta_est")))
   if (component(object, "reml") != 1) {
     stop("Kenward-Roger is only for REML")
@@ -60,7 +60,7 @@ h_df_md_kr <- function(object, contrast, linear = FALSE) {
     num_df = nrow(contrast),
     denom_df = df$m,
     f_stat = f_star[1, 1],
-    p_val = pf(f_star[1, 1], nrow(contrast), df$m, lower.tail = FALSE)
+    p_val = stats::pf(f_star[1, 1], nrow(contrast), df$m, lower.tail = FALSE)
   )
 }
 
@@ -79,7 +79,7 @@ h_df_md_kr <- function(object, contrast, linear = FALSE) {
 #' @keywords internal
 h_df_1d_kr <- function(object, contrast, linear = FALSE) {
   assert_class(object, "mmrm")
-  assert_logical(linear)
+  assert_flag(linear)
   assert_numeric(contrast, len = length(component(object, "beta_est")))
   if (component(object, "reml") != 1) {
     stop("Kenward-Roger is only for REML!")
@@ -181,7 +181,7 @@ h_kr_df <- function(v0, l, w, p) {
 #'
 #' @keywords internal
 h_var_adj <- function(v, w, p, q, r, linear = FALSE) {
-  assert_logical(linear)
+  assert_flag(linear)
   n_beta <- ncol(v)
   assert_matrix(v, nrows = n_beta)
   n_theta <- ncol(w)